+++ Tulpep.PayULibrary/Models/Response/Response_Cross/Response_PMNETWORKVALUE.cs
@@ -2,7 +2,7 @@
 {
     public class Response_PMNETWORKVALUE
     {
-        public double value { get; set; }
+        public decimal value { get; set; }
         public string currency { get; set; }
     }
 }
